refactor(confirmations): share amount formatting utility
