+++ ui/pages/confirmations/components/confirm/info/hooks/useGasFeeToken.ts
@@ -18,7 +18,7 @@ import {
   getShouldShowFiat,
   selectTransactionAvailableBalance,
 } from '../../../../../../selectors';
-import { formatAmount } from '../../../simulation-details/formatAmount';
+import { formatAmount } from '../../../../../../../shared/lib/format-amount';
 import { useConfirmContext } from '../../../../context/confirm';
 import { useFeeCalculations } from './useFeeCalculations';
 import { useNativeCurrencySymbol } from './useNativeCurrencySymbol';
