+++ ui/pages/confirmations/components/confirm/info/shared/gas-fee-token-list-item/gas-fee-token-list-item.tsx
@@ -30,7 +30,7 @@ import { useI18nContext } from '../../../../../../../hooks/useI18nContext';
 import { useGasFeeToken } from '../../hooks/useGasFeeToken';
 import { getCurrentCurrency } from '../../../../../../../ducks/metamask/metamask';
 import { GasFeeTokenIcon, GasFeeTokenIconSize } from '../gas-fee-token-icon';
-import { formatAmount } from '../../../../simulation-details/formatAmount';
+import { formatAmount } from '../../../../../../../../shared/lib/format-amount';
 import { getIntlLocale } from '../../../../../../../ducks/locale/locale';
 
 export type GasFeeTokenListItemProps = {
