+++ ui/pages/confirmations/components/simulation-details/amount-pill.tsx
@@ -16,8 +16,8 @@ import Tooltip from '../../../../components/ui/tooltip';
 import { getIntlLocale } from '../../../../ducks/locale/locale';
 import { shortenString as shortenAssetId } from '../../../../helpers/utils/util';
 import { useI18nContext } from '../../../../hooks/useI18nContext';
+import { formatAmount, formatAmountMaxPrecision } from '../../../../../shared/lib/format-amount';
 import { AssetIdentifier } from './types';
-import { formatAmount, formatAmountMaxPrecision } from './formatAmount';
 
 /**
  * Displays a pill with an amount and a background color indicating whether the amount
